Move ConfigProvider to provider subpackage
Also move FileConfigProviderTest to provider subpackage
