+++ clients/src/main/java/org/apache/kafka/common/config/ConfigTransformerResult.java
@@ -16,6 +16,8 @@
  */
 package org.apache.kafka.common.config;
 
+import org.apache.kafka.common.config.provider.ConfigProvider;
+
 import java.util.Map;
 
 /**
