+++ clients/src/main/java/org/apache/kafka/common/config/provider/ConfigProvider.java
@@ -14,9 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.config;
+package org.apache.kafka.common.config.provider;
 
 import org.apache.kafka.common.Configurable;
+import org.apache.kafka.common.config.ConfigChangeCallback;
+import org.apache.kafka.common.config.ConfigData;
 
 import java.io.Closeable;
 import java.util.Set;
