+++ clients/src/main/java/org/apache/kafka/common/config/provider/FileConfigProvider.java
@@ -18,7 +18,6 @@ package org.apache.kafka.common.config.provider;
 
 import org.apache.kafka.common.config.ConfigData;
 import org.apache.kafka.common.config.ConfigException;
-import org.apache.kafka.common.config.ConfigProvider;
 
 import java.io.BufferedReader;
 import java.io.FileInputStream;
