+++ clients/src/main/resources/META-INF/services/org.apache.kafka.common.config.provider.ConfigProvider
