+++ clients/src/test/java/org/apache/kafka/common/config/provider/FileConfigProviderTest.java
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.config;
+package org.apache.kafka.common.config.provider;
 
-import org.apache.kafka.common.config.provider.FileConfigProvider;
+import org.apache.kafka.common.config.ConfigData;
 import org.junit.Before;
 import org.junit.Test;
