+++ connect/runtime/src/main/java/org/apache/kafka/connect/runtime/distributed/ClusterConfigState.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.connect.runtime.distributed;
 
+import org.apache.kafka.common.config.provider.ConfigProvider;
 import org.apache.kafka.connect.runtime.WorkerConfigTransformer;
 import org.apache.kafka.connect.runtime.TargetState;
 import org.apache.kafka.connect.util.ConnectorTaskId;
@@ -110,7 +111,7 @@ public class ClusterConfigState {
      * Get the configuration for a connector.  The configuration will have been transformed by
      * {@link org.apache.kafka.common.config.ConfigTransformer} by having all variable
      * references replaced with the current values from external instances of
-     * {@link org.apache.kafka.common.config.ConfigProvider}, and may include secrets.
+     * {@link ConfigProvider}, and may include secrets.
      * @param connector name of the connector
      * @return a map containing configuration parameters
      */
@@ -135,7 +136,7 @@ public class ClusterConfigState {
      * Get the configuration for a task.  The configuration will have been transformed by
      * {@link org.apache.kafka.common.config.ConfigTransformer} by having all variable
      * references replaced with the current values from external instances of
-     * {@link org.apache.kafka.common.config.ConfigProvider}, and may include secrets.
+     * {@link ConfigProvider}, and may include secrets.
      * @param task id of the task
      * @return a map containing configuration parameters
      */
